requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-filter==24.3

psycopg2-binary==2.9.10

asgiref==3.8.1
sqlparse==0.5.3
python-dotenv==1.0.0
markdown==3.7
requests==2.32.3